+++ web/static/react/assets/PracticeProfilePage-CwLumlmP.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-BqsLoW2D.js";import{t}from"./LegalSkillsProfilePage-DyCd_zzy.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as PracticeProfilePage};